Move create-empty-results and send-challenge-list-email to jobs/utils
