+++ spec/filter-manage-actions-spec.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-undef */
 /* filterManageActions should filter out all actions for a spotId if the last action is 'Closed Spot' */
-const { filterManageActions } = require('../jobs').sendChallengeListEmail;
+const { filterManageActions } = require('../jobs/utils').sendChallengeListEmail;
 
 console.log('==> FILTER MANAGE ACTIONS');
 describe('filterManageActions', () => {
